+++ changelog.md
@@ -6,6 +6,7 @@
 > This version is not released yet and is under active development.
 
 - Add support for auto-conversion of GitHub alerts into MyST admonitions in Sphinx extension.
+- Rename `click-extra` demo CLI to `click-extra-demo` to avoid confusion with the package name.
 - Run tests on `ubuntu-slim` GitHub Actions runner.
 - Run docs update job on `ubuntu-slim` runner.